Update README.md
In consideration with the above-mentioned problems the solution has been proposed. One of the main advantages of electric bike is that they use limited non-renewable fuel and don't pump out a lot of harmful emissions. The frame is designed to carry a load of 300kg, telescopic type suspension, design of shaft, sprocket, weight distribution and carried out structural analysis of frame and suspension for the objective of a comfortable, compact, high speed and efficient bike.
The electric bike's most important feature is that it is simple in construction due to the use of lesser components and it produces no pollution, is environmental friendly, and operates quietly.